implement get all restcontroller

+++ src/main/java/com/example/libreria/infraestructure/crud_interface/AutorCrudRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.example.libreria.infraestructure.entity.Autor;
 
-public interface AutorCrudRepository extends CrudRepository<Autor, Integer> {
+public interface AutorCrudRepository extends CrudRepository<Autor, Long> {
     List<Autor> findByIdAutor(Long idAutor);
 }
